pyproject.toml
[project]
name = "langchain-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.14",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
]
